requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.3
pytest>=8.3
pytest-django>=4.8
django-environ>=0.10.0
drf-standardized-errors